feat: rename artifact
