--- executions/graphql-kotlin-transaction-batcher-instrumentation-level/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/level/execution/ExecutionLevelInstrumentation.kt
@@ -1,9 +1,0 @@
-package com.expediagroup.graphql.transactionbatcher.instrumentation.level.execution
-
-import graphql.execution.instrumentation.Instrumentation
-
-interface ExecutionLevelInstrumentation : Instrumentation {
-    fun beginExecutionLevel(
-        parameters: ExecutionLevelInstrumentationParameters
-    ): ExecutionLevelInstrumentationContext
-}
