--- executions/graphql-kotlin-transaction-batcher-instrumentation-level/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/level/execution/ExecutionLevelInstrumentationContext.kt
@@ -1,7 +1,0 @@
-package com.expediagroup.graphql.transactionbatcher.instrumentation.level.execution
-
-import com.expediagroup.graphql.transactionbatcher.instrumentation.level.state.Level
-
-interface ExecutionLevelInstrumentationContext {
-    fun onLevelDispatched(level: Level)
-}
