--- executions/graphql-kotlin-transaction-batcher-instrumentation-level/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/level/execution/ExecutionLevelInstrumentationParameters.kt
@@ -1,7 +1,0 @@
-package com.expediagroup.graphql.transactionbatcher.instrumentation.level.execution
-
-import graphql.execution.ExecutionContext
-
-data class ExecutionLevelInstrumentationParameters(
-    val executionContext: ExecutionContext
-)
