--- executions/graphql-kotlin-transaction-batcher-instrumentation-level/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/level/extensions/ConcurrentHashMapExtensions.kt
@@ -1,10 +1,0 @@
-package com.expediagroup.graphql.transactionbatcher.instrumentation.level.extensions
-
-import java.util.concurrent.ConcurrentHashMap
-
-fun <K, V, R> ConcurrentHashMap<K, V>.synchronizeIfPresent(key: K, block: (V) -> R): R? =
-    this[key]?.let { value ->
-        synchronized(value) {
-            block(value)
-        }
-    }
