--- executions/graphql-kotlin-transaction-batcher-instrumentation-level/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/level/extensions/FieldValueInfosExtensions.kt
@@ -1,14 +1,0 @@
-package com.expediagroup.graphql.transactionbatcher.instrumentation.level.extensions
-
-import graphql.execution.FieldValueInfo
-
-fun List<FieldValueInfo>.getExpectedStrategyCalls(): Int {
-    var count = 0
-    this.forEach { fieldValueInfo ->
-        if (fieldValueInfo.completeValueType == FieldValueInfo.CompleteValueType.OBJECT)
-            count++
-        else if (fieldValueInfo.completeValueType == FieldValueInfo.CompleteValueType.LIST)
-            count += fieldValueInfo.fieldValueInfos.getExpectedStrategyCalls()
-    }
-    return count
-}
