--- executions/graphql-kotlin-transaction-batcher-instrumentation-level/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/level/state/Level.kt
@@ -1,7 +1,0 @@
-package com.expediagroup.graphql.transactionbatcher.instrumentation.level.state
-
-data class Level(val number: Int) {
-    fun next(): Level = Level(number + 1)
-    fun previous(): Level = Level(number - 1)
-    fun isFirst(): Boolean = number == 1
-}
