+++ executions/graphql-kotlin-transaction-batcher-instrumentation/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/execution/ExecutionLevelInstrumentation.kt
@@ -1,0 +1,9 @@
+package com.expediagroup.graphql.transactionbatcher.instrumentation.execution
+
+import graphql.execution.instrumentation.Instrumentation
+
+interface ExecutionLevelInstrumentation : Instrumentation {
+    fun beginExecutionLevel(
+        parameters: ExecutionLevelInstrumentationParameters
+    ): ExecutionLevelInstrumentationContext
+}
