+++ executions/graphql-kotlin-transaction-batcher-instrumentation/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/execution/ExecutionLevelInstrumentationContext.kt
@@ -1,0 +1,7 @@
+package com.expediagroup.graphql.transactionbatcher.instrumentation.execution
+
+import com.expediagroup.graphql.transactionbatcher.instrumentation.state.Level
+
+interface ExecutionLevelInstrumentationContext {
+    fun onLevelDispatched(level: Level)
+}
