+++ executions/graphql-kotlin-transaction-batcher-instrumentation/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/execution/ExecutionLevelInstrumentationParameters.kt
@@ -1,0 +1,7 @@
+package com.expediagroup.graphql.transactionbatcher.instrumentation.execution
+
+import graphql.execution.ExecutionContext
+
+data class ExecutionLevelInstrumentationParameters(
+    val executionContext: ExecutionContext
+)
