+++ executions/graphql-kotlin-transaction-batcher-instrumentation/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/extensions/ConcurrentHashMapExtensions.kt
@@ -1,0 +1,10 @@
+package com.expediagroup.graphql.transactionbatcher.instrumentation.extensions
+
+import java.util.concurrent.ConcurrentHashMap
+
+fun <K, V, R> ConcurrentHashMap<K, V>.synchronizeIfPresent(key: K, block: (V) -> R): R? =
+    this[key]?.let { value ->
+        synchronized(value) {
+            block(value)
+        }
+    }
