+++ executions/graphql-kotlin-transaction-batcher-instrumentation/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/extensions/ExecutionContextExtensions.kt
@@ -1,0 +1,23 @@
+package com.expediagroup.graphql.transactionbatcher.instrumentation.extensions
+
+import graphql.analysis.QueryTraverser
+import graphql.analysis.QueryVisitorFieldEnvironment
+import graphql.execution.ExecutionContext
+import kotlin.math.max
+
+fun ExecutionContext.getDocumentHeight(): Int {
+    val getFieldDepth: (QueryVisitorFieldEnvironment?) -> Int = { queryVisitor ->
+        var hasQueryVisitor = queryVisitor
+        var height = 1
+        while (hasQueryVisitor != null) {
+            hasQueryVisitor = hasQueryVisitor.parentEnvironment
+            height++
+        }
+        height
+    }
+    return QueryTraverser.Builder().schema(graphQLSchema).document(document).variables(variables).build()
+        .reducePreOrder(
+            { queryVisitor, height -> max(getFieldDepth(queryVisitor.parentEnvironment), height) },
+            0
+        )
+}
