+++ executions/graphql-kotlin-transaction-batcher-instrumentation/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/extensions/FieldValueInfosExtensions.kt
@@ -1,0 +1,14 @@
+package com.expediagroup.graphql.transactionbatcher.instrumentation.extensions
+
+import graphql.execution.FieldValueInfo
+
+fun List<FieldValueInfo>.getExpectedStrategyCalls(): Int {
+    var count = 0
+    this.forEach { fieldValueInfo ->
+        if (fieldValueInfo.completeValueType == FieldValueInfo.CompleteValueType.OBJECT)
+            count++
+        else if (fieldValueInfo.completeValueType == FieldValueInfo.CompleteValueType.LIST)
+            count += fieldValueInfo.fieldValueInfos.getExpectedStrategyCalls()
+    }
+    return count
+}
