+++ executions/graphql-kotlin-transaction-batcher-instrumentation/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/state/Level.kt
@@ -1,0 +1,7 @@
+package com.expediagroup.graphql.transactionbatcher.instrumentation.state
+
+data class Level(val number: Int) {
+    fun next(): Level = Level(number + 1)
+    fun previous(): Level = Level(number - 1)
+    fun isFirst(): Boolean = number == 1
+}
